feat(molecules): create unit testing for SummaryStep and move it to molecules folder
