+++ src/components/molecules/SummaryStep/index.tsx
@@ -1,19 +1,14 @@
 import React from "react";
 import Link from "next/link";
-
-interface SummaryStepProps {
-  collectedData: {
-    email: string;
-    age: number;
-  };
-}
+import { SummaryStepProps } from "@/components/types";
 
 const SummaryStep: React.FC<SummaryStepProps> = (props) => {
+  const {collectedData:{age,email}}=props
   return (
     <div className="flex flex-col gap-4 w-full">
       <div className="flex flex-col gap-2 items-start">
-        <div>Email: {props.collectedData.email}</div>
-        <div>Age: {props.collectedData.age}</div>
+        <div>Email: {email}</div>
+        <div>Age: {age}</div>
       </div>
       <Link className="btn btn-primary" href="/purchased=dev_ins">
         Purchase
